build: Commit Functions for Client Operations
Accepted clients will be added into a client pool data structure, where
the each item in the pool is a `client_t` data structure that contains
the client's user ID, their name, and the file descriptor of the
accepted connection.

Additionally, to broadcast a message, the server will iterate over the
client pool and send the message to all clients except for the sender.